+++ src/holocron/providers/__init__.py
@@ -1,5 +1,6 @@
 from .gitlab import GitLabProvider
 from .github import GitHubProvider
+from .azure import AzureDevOpsProvider
 from .base import Provider
 
-__all__ = ["GitLabProvider", "GitHubProvider", "Provider"]
+__all__ = ["GitLabProvider", "GitHubProvider", "AzureDevOpsProvider", "Provider"]